Move add_variant_to_checkout to tests utils
