+++ saleor/graphql/checkout/tests/mutations/test_checkout_lines_update.py
@@ -17,8 +17,8 @@ from .....checkout.calculations import (
 from .....checkout.error_codes import CheckoutErrorCode
 from .....checkout.fetch import fetch_checkout_info, fetch_checkout_lines
 from .....checkout.models import Checkout, CheckoutLine
+from .....checkout.tests.utils import add_variant_to_checkout
 from .....checkout.utils import (
-    add_variant_to_checkout,
     add_variants_to_checkout,
     calculate_checkout_quantity,
     invalidate_checkout,
