+++ saleor/plugins/avatax/tests/test_avatax.py
@@ -14,7 +14,7 @@ from requests_hardened import HTTPSession
 
 from ....account.models import Address
 from ....checkout.fetch import CheckoutInfo, fetch_checkout_info, fetch_checkout_lines
-from ....checkout.utils import add_variant_to_checkout
+from ....checkout.tests.utils import add_variant_to_checkout
 from ....core.prices import quantize_price
 from ....core.taxes import (
     TaxDataErrorMessage,
